+++ command/packet_screen.go
@@ -29,7 +29,7 @@ func (packet ScreenPacket) Read(c *Client) error {
 
 	rlen, err := io.ReadFull(c, c.Screen.Buffer)
 
-	if int32(rlen) != len {
+	if rlen != len {
 		return errors.New("EOF")
 	}
